test: move pr509–pr531 lowering seams under test/lowering/
- Relocate helper and integration tests; fix imports and fixture paths
- Update test/README index and subsystem blurb
- Point archive design doc at new pr509 ea materialization test path

Fixes #1080

Made-with: Cursor
