+++ test/lowering/pr510_op_expansion_orchestration_helpers.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 
-import { DiagnosticIds, type Diagnostic } from '../src/diagnosticTypes.js';
-import type { AsmInstructionNode, AsmOperandNode, OpDeclNode, SourceSpan } from '../src/frontend/ast.js';
-import { createOpExpansionOrchestrationHelpers } from '../src/lowering/opExpansionOrchestration.js';
+import { DiagnosticIds, type Diagnostic } from '../../src/diagnosticTypes.js';
+import type { AsmInstructionNode, AsmOperandNode, OpDeclNode, SourceSpan } from '../../src/frontend/ast.js';
+import { createOpExpansionOrchestrationHelpers } from '../../src/lowering/opExpansionOrchestration.js';
 
 const span: SourceSpan = {
   file: 'test.zax',
