+++ test/lowering/pr510_op_substitution_helpers.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 
-import { DiagnosticIds, type Diagnostic } from '../src/diagnosticTypes.js';
-import type { AsmOperandNode, EaExprNode, ImmExprNode, SourceSpan } from '../src/frontend/ast.js';
-import { createOpSubstitutionHelpers } from '../src/lowering/opSubstitution.js';
+import { DiagnosticIds, type Diagnostic } from '../../src/diagnosticTypes.js';
+import type { AsmOperandNode, EaExprNode, ImmExprNode, SourceSpan } from '../../src/frontend/ast.js';
+import { createOpSubstitutionHelpers } from '../../src/lowering/opSubstitution.js';
 
 const span: SourceSpan = {
   file: 'test.zax',
